requirements.txt
openai-whisper
gradio
opencc-python-reimplemented